feat: 존재하지 않는 여행기 삭제시 400을 응답하는 기능 구현

+++ backend/src/main/java/kr/touroot/travelogue/service/TravelogueService.java
@@ -38,6 +38,10 @@ public class TravelogueService {
     }
 
     public void deleteById(Long id) {
+        if (!travelogueRepository.existsById(id)) {
+            throw new BadRequestException("존재하지 않는 여행기입니다.");
+        }
+
         travelogueRepository.deleteById(id);
     }
 }
